requirements.txt
cryptography==2.4.2
Django==2.0
numpy==1.15.4
lxml==4.2.5
Cython==0.29.6
pyOpenSSL==19.0.0
protobuf==3.6.1
ipaddr==2.2.0
django_mptt==0.9.1
django-cleanup==3.2.0
ordereddict==1.1
usercustomize==1.0.0
wincertstore==0.2
Pillow==5.4.1
django-test-without-migrations==0.6
django-nose==1.4.6
mysqlclient==1.4.2.post1
